Kindle Image Server/requirements.txt
flask
jinja2
playwright
requests
Pillow
async_lru
